+++ Jibberwock.Persistence.DataAccess/Commands/Security/DeleteAccessControlEntry.cs
@@ -27,7 +27,7 @@ namespace Jibberwock.Persistence.DataAccess.Commands.Security
         [Required]
         public AccessControlEntry AccessControlEntry { get; set; }
 
-        public DeleteAccessControlEntry(ILogger logger, User performedBy, string connectionId, int serviceId, string comment, AccessControlEntry accessControlEntry)
+        public DeleteAccessControlEntry(ILogger logger, User performedBy, string connectionId, long serviceId, string comment, AccessControlEntry accessControlEntry)
             : base(logger, performedBy, connectionId, serviceId, comment)
         {
             AccessControlEntry = accessControlEntry;
